Display OwnEmailServiceNote component on the second part of the MSS step
Before this component was displayed only on the first part of the MSS
step.

[MAILPOET-4818]
