+++ mailpoet/assets/js/src/wizard/steps/pitch_mss_step.tsx
@@ -24,7 +24,12 @@ function WelcomeWizardPitchMSSStep({
         />
       );
     case 'second':
-      return <MSSStepSecondPart setStepPart={setStepPart} />;
+      return (
+        <MSSStepSecondPart
+          setStepPart={setStepPart}
+          finishWizard={finishWizard}
+        />
+      );
     default:
       return <MSSStepThirdPart finishWizard={finishWizard} />;
   }
