+++ mailpoet/assets/js/src/wizard/steps/pitch_mss_step/first_part.tsx
@@ -1,81 +1,9 @@
 import { external, Icon } from '@wordpress/icons';
-import { useState } from '@wordpress/element';
-import { Modal } from '@wordpress/components';
-import ReactStringReplace from 'react-string-replace';
 import { Heading } from 'common/typography/heading/heading';
 import { MailPoet } from 'mailpoet';
 import { List } from 'common/typography/list/list';
 import { Button } from 'common';
-
-type OwnEmailDeliveryServicePropType = {
-  finishWizard: (redirect_url?: string) => void;
-};
-
-function OwnEmailDeliveryService({
-  finishWizard,
-}: OwnEmailDeliveryServicePropType): JSX.Element {
-  const [confirmationModalIsOpen, setConfirmationModalOpen] = useState(false);
-  const openConfirmationModal = (e) => {
-    e.preventDefault();
-    setConfirmationModalOpen(true);
-  };
-  const closeConfirmationModal = () => setConfirmationModalOpen(false);
-  const finishWithOwnService = (e) => {
-    e.preventDefault();
-    finishWizard('admin.php?page=mailpoet-settings#/mta/other');
-  };
-
-  return (
-    <>
-      <p>
-        {ReactStringReplace(
-          MailPoet.I18n.t('welcomeWizardMSSAdvancedUsers'),
-          /\[link](.*?)\[\/link]/g,
-          (match, index) => (
-            <a
-              key={index}
-              className="mailpoet-link"
-              onClick={openConfirmationModal}
-              href="#"
-            >
-              {match}
-            </a>
-          ),
-        )}
-      </p>
-      {confirmationModalIsOpen && (
-        <Modal
-          className="mailpoet-welcome-wizard-confirmation-modal"
-          title={MailPoet.I18n.t('welcomeWizardMSSConfirmationModalTitle')}
-          onRequestClose={closeConfirmationModal}
-        >
-          <p>
-            {window.mailpoet_mail_function_enabled
-              ? MailPoet.I18n.t(
-                  'welcomeWizardMSSConfirmationModalFirstParagraph',
-                )
-              : MailPoet.I18n.t(
-                  'welcomeWizardMSSConfirmationModalFirstParagraphWithoutMailFunction',
-                )}
-          </p>
-          <p>
-            {MailPoet.I18n.t(
-              'welcomeWizardMSSConfirmationModalSecondParagraph',
-            )}
-          </p>
-          <div className="mailpoet-welcome-wizard-confirmation-modal-buttons">
-            <Button variant="secondary" onClick={closeConfirmationModal}>
-              {MailPoet.I18n.t('welcomeWizardMSSConfirmationModalGoBackButton')}
-            </Button>
-            <Button onClick={finishWithOwnService}>
-              {MailPoet.I18n.t('welcomeWizardMSSConfirmationModalOkButton')}
-            </Button>
-          </div>
-        </Modal>
-      )}
-    </>
-  );
-}
+import { OwnEmailServiceNote } from './own_email_service_note';
 
 const mailpoetAccountUrl = MailPoet.MailPoetComUrlFactory.getPurchasePlanUrl(
   MailPoet.subscribersCount,
@@ -139,7 +67,7 @@ function MSSStepFirstPart({
       <div className="mailpoet-gap" />
       <div className="mailpoet-gap" />
 
-      <OwnEmailDeliveryService finishWizard={finishWizard} />
+      <OwnEmailServiceNote finishWizard={finishWizard} />
     </>
   );
 }
